+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+
+- Support for the [new Apple Podcast categories][category-new-2019] that were [added August 9th 2019][category-published-2019].
+- Documentation of the Warning classes defined by PodGen.
+
+[category-new-2019]: https://podnews.net/article/apple-changed-podcast-categories-2019
+[category-published-2019]: https://itunespartner.apple.com/podcasts/whats-new/100002598
+
+### Changed
+
+- Using one of the old iTunes (sub)categories will now generate a `LegacyCategoryWarning`.
+
+### Deprecated
+
+- Importing `NotSupportedByItunesWarning` from `podgen.not_supported_by_itunes_warning`. Import from `podgen` instead.
+
 
 ## [1.0.1] - 2019-10-12
 ### Added